Extract filtering from Sanitize to PostSanitizeFilter
