+++ backend/src/main/kotlin/app/ehrenamtskarte/backend/stores/importer/steps/PostSanitizeFilter.kt
@@ -1,0 +1,36 @@
+package app.ehrenamtskarte.backend.stores.importer.steps
+
+import app.ehrenamtskarte.backend.stores.geocoding.FeatureFetcher
+import app.ehrenamtskarte.backend.common.STATE
+import app.ehrenamtskarte.backend.stores.geocoding.isInBoundingBox
+import app.ehrenamtskarte.backend.stores.importer.PipelineStep
+import app.ehrenamtskarte.backend.stores.importer.types.AcceptingStore
+import io.ktor.client.*
+import kotlinx.coroutines.runBlocking
+import org.slf4j.Logger
+
+class PostSanitizeFilter(private val logger: Logger, httpClient: HttpClient): PipelineStep<List<AcceptingStore>, List<AcceptingStore>>() {
+    private val featureFetcher = FeatureFetcher(httpClient)
+
+    override fun execute(input: List<AcceptingStore>): List<AcceptingStore> = runBlocking {
+        val stateBbox = featureFetcher.queryFeatures(listOf(Pair("state", STATE))).first().bbox
+
+        input.filter {
+            if (it.longitude == null || it.latitude == null) {
+                logger.info("'${it.name}, ${it.location}' was filtered out because longitude or latitude are null")
+                return@filter false
+            }
+            if (!it.isInBoundingBox(stateBbox)) {
+                logger.info("'${it.name}, ${it.location}' was filtered out because it is outside of $STATE")
+                return@filter false
+            }
+            if (it.postalCode == null) {
+                // Probably because it is outside of the state but inside the bounding box of the state
+                logger.info("'${it.name}, ${it.location}' was filtered out because its postal code is null")
+                return@filter false
+            }
+            true
+        }
+    }
+
+}
